requirements.txt
# Multi-Microservice Pattern Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis and Messaging
redis[hiredis]==5.0.4

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Validation and Serialization
email-validator==2.1.0
phonenumbers==8.13.26

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
psutil==5.9.5

# Distributed Tracing
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-aiohttp-client==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-asyncpg==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-proto==1.24.0
grpcio==1.60.0

# Development and Testing
setuptools<81
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0

# Performance
orjson==3.9.10
ujson==5.8.0

# Time and Date
python-dateutil==2.8.2
pytz==2023.3

# Async utilities
asyncio-mqtt==0.13.0
aiofiles==23.2.1

# Azure Storage
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
